interfaces added for json formats
